requirements.txt
django>=1.8.5
htmlmin>=0.1.10
markdown>=2.6.2
dirsync>=2.2.2
django-import-export